refactor : fix the parameter int to Long

+++ src/main/java/calculator/view/ResultView.java
@@ -1,7 +1,7 @@
 package calculator.view;
 
 public class ResultView {
-    public static void promptForAdditionResult(int result) {
+    public static void promptForAdditionResult(Long result) {
         System.out.print("결과 : " + result);
     }
 }
